feat(api): new api style, user context, utils
